requirements.txt
Django==5.2.6
psycopg[binary]==3.2.10
mypy==1.18.2
django-stubs==5.2.5
gunicorn==23.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
drf-spectacular==0.28.0
mailersend==2.0.0
